플레이어 카메라 컬링은 정상적으로 수행 되나, 그림자에서 오류 발생. fps 저하로 인해 더 진행하지 않기로 결정

+++ DX12App/Shaders/common.hlsl
@@ -55,7 +55,7 @@ cbuffer StandardCB : register(b5)
     float gZSplit0 : packoffset(c0.x);
     float gZSplit1 : packoffset(c0.y);
     float gZSplit2 : packoffset(c0.z);
-    int gInstancingOffset : packoffset(c0.w);
+    float gInstancingOffset : packoffset(c0.w);
 }
 
 struct InstancingInfo
